fix(core/graph-context): correctly index signed constructs with multiple signers
The svm addon introduced constructs with multiple signers. When the graph context indexes signers and the upstream dependencies of their downstream dependencies, it was supposed to filter out the signers. However, only the one signer we were looking at was filtered. With this change, all signers are filtered out, preventing some weird results when re-executing a state-managed runbook.